Version 0.15.0 Alpha.4 [Prerelease]
New features:
- VPtr
- AuxECS

Bugfixes:
- Calligraphics font texture size option fixes
- Rare Unika source generator failures

+++ Calligraphics/Internal/Harfbuzz/Bitmap/SdfRasterizer.cs
@@ -32,7 +32,7 @@ namespace Latios.Calligraphics.HarfBuzz.Bitmap
                 var distSq         = distSqs[i];
                 var signedDistance = sign * math.min(math.sqrt(distSq) / spread, 1f);  // in [-1, 1] range
                 var scaled         = (signedDistance + 1f) * 65535f / 2f;  // unorm correction
-                buffer[i]          = (byte)scaled;
+                buffer[i]          = (ushort)scaled;
             }
         }
